Propagate sign-out and database failures

+++ feature/diary-profile/src/commonMain/kotlin/com/foreverrafs/superdiary/profile/domain/usecase/SignOutUseCase.kt
@@ -11,7 +11,11 @@ class SignOutUseCase(
 ) {
     suspend operator fun invoke(): Result<Unit> {
         return try {
-            authApi.signOut()
+            val signOutResult = authApi.signOut()
+            if (signOutResult.isFailure) {
+                return signOutResult
+            }
+
             preferences.clear()
             dataSource.deleteAll()
             dataSource.clearChatMessages()
